Ârefactor: Creating github conteroller, and user & pull-Request repositories
